Начало работы над тестами залов
