+++ phpunit.xml
@@ -21,14 +21,10 @@
         </include>
     </coverage>
     <php>
-<<<<<<< HEAD
-=======
-
->>>>>>> AnatolyKonakovTests
         <env name="APP_ENV" value="testing"/>
         <env name="BCRYPT_ROUNDS" value="4"/>
         <env name="CACHE_DRIVER" value="array"/>
-        <env name="DB_DATABASE" value="homestead"/>
+        <env name="DB_DATABASE" value="testing"/>
         <env name="MAIL_MAILER" value="array"/>
         <env name="QUEUE_CONNECTION" value="sync"/>
         <env name="SESSION_DRIVER" value="array"/>
