+++ tests/routList.txt
@@ -2,7 +2,7 @@ DELETE          account/profiles/{profile} account.profiles.destroy › Accou…
 DELETE          account/users/{user} account.users.destroy › Account\UserCo…   Не прописан в самом контроллере
 PUT|PATCH       account/profiles/{profile} account.profiles.update › Accoun…
 
- 
+
   POST            email/verification-notification verification.send › Auth\EmailVerificationNotificationController@store
   POST            subscriptions ............................... subscriptions.store › SubscriptionController@store
   PUT|PATCH       subscriptions/{subscription} .............. subscriptions.update › SubscriptionController@update
@@ -21,6 +21,19 @@ PUT|PATCH       account/profiles/{profile} account.profiles.update › Accoun…
   POST            reset-password .............................. password.update › Auth\NewPasswordController@store
   GET|HEAD        reset-password/{token} ...................... password.reset › Auth\NewPasswordController@create
   GET|HEAD        sanctum/csrf-cookie .......... sanctum.csrf-cookie › Laravel\Sanctum › CsrfCookieController@show
+
+
+
+  GET|HEAD        gym/{id}/{city_id} ................................................................. gyms.show › GymController@show
+  GET|HEAD        gymReviews ........................................................... gymReviews.index › GymReviewController@index
+  POST            gymReviews ........................................................... gymReviews.store › GymReviewController@store
+  GET|HEAD        gymReviews/create .................................................. gymReviews.create › GymReviewController@create
+  GET|HEAD        gymReviews/{gymReview} ................................................. gymReviews.show › GymReviewController@show
+  PUT|PATCH       gymReviews/{gymReview} ............................................. gymReviews.update › GymReviewController@update
+  DELETE          gymReviews/{gymReview} ........................................... gymReviews.destroy › GymReviewController@destroy
+  GET|HEAD        gymReviews/{gymReview}/edit ............................................ gymReviews.edit › GymReviewController@edit
+  GET|HEAD        gyms/{city_id} ................................................................... gyms.index › GymController@index
+
